Adds a few more get and post endpoints
Now if you go to /date it returns the current date, and /count/<number> it returns the count passed in as json. You can also make a post request to /user-info with json including a name string and an age int and it will return a string with those parameters interpolated. Also, some of the code not being used was removed.